requirements.txt
llama-index==0.10.50
aiohttp>=3.8.6,<4.0.0
charset-normalizer
colorama
requests
pillow
pydantic
supabase
markdownify==0.12.1
magika
scikit-learn
torch
torchvision
torchaudio
docx2txt
python-pptx
Pyarrow
poppler-utils
python-magic
pandas
pytesseract
cssutils
beautifulsoup4
PyMuPDF
pdf2image
python-dotenv
git+https://github.com/openai/whisper.git
moviepy==1.0.3
openpyxl
pytube
ffmpeg-python
python-docx
pymupdf4llm
pytest-playwright
playwright
jwt
sentence-transformers
marker-pdf
modal